+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageListHelperViews.swift
@@ -124,6 +124,7 @@ public struct MessageReadIndicatorView: View {
             .customizable()
             .foregroundColor(!readUsers.isEmpty ? colors.tintColor : Color(colors.textLowEmphasis))
             .frame(height: 16)
+            .opacity(localState == .sendingFailed ? 0.0 : 1)
             .accessibilityLabel(
                 Text(
                     readUsers.isEmpty ? L10n.Message.ReadStatus.seenByNoOne : L10n.Message.ReadStatus.seenByOthers
@@ -136,7 +137,12 @@ public struct MessageReadIndicatorView: View {
     }
     
     private var image: UIImage {
-        !readUsers.isEmpty ? images.readByAll : (localState == .pendingSend ? images.messageReceiptSending : images.messageSent)
+        !readUsers.isEmpty ? images.readByAll : (isMessageSending ? images.messageReceiptSending : images.messageSent)
+    }
+
+    private var isMessageSending: Bool {
+        localState == .sending
+            || localState == .pendingSend
     }
 }